change type for labels

+++ src/labelingcc.cpp
@@ -55,23 +55,23 @@ int main( int narg, char **argv ) {
 	
 	QString input( inputFileName.c_str() ) ;
 	QString output( outputFileName.c_str() ) ;
-	BillonTpl<arma::u16> *labels ;
+	BillonTpl<arma::u32> *labels ;
 	if ( vm.count ( "input" ) ) {
 		Pgm3dFactory<arma::u8> factory ;
 		BillonTpl<arma::u8> * skel = factory.read( input ) ;
 		factory.correctEncoding( skel ) ;
-		ConnexComponentExtractor<arma::u8,arma::u16> CCE ;
+		ConnexComponentExtractor<arma::u8,arma::u32> CCE ;
 		labels = CCE.run( *skel ) ;
 		delete skel ;
 	} else {
-		Pgm3dFactory<arma::u16> factory ;
+		Pgm3dFactory<arma::u32> factory ;
 		labels = factory.read( input ) ;
 	}
 
-	GrayLevelHistogram<arma::u16> histogram( *labels ) ;
-	std::list< std::pair<arma::u16,int32_t> > sizeCC ;
-	std::list< std::pair<arma::u16,int32_t> >::iterator pos ;
-	for ( GrayLevelHistogram<arma::u16>::THistogram::const_iterator it=histogram._bin.begin(); it != histogram._bin.end();it++ ) {
+	GrayLevelHistogram<arma::u32> histogram( *labels ) ;
+	std::list< std::pair<arma::u32,int32_t> > sizeCC ;
+	std::list< std::pair<arma::u32,int32_t> >::iterator pos ;
+	for ( GrayLevelHistogram<arma::u32>::THistogram::const_iterator it=histogram._bin.begin(); it != histogram._bin.end();it++ ) {
 		std::cout<<"Key : "<<it->first<<" Frequency : "<<it->second<<std::endl;
 
 		pos = sizeCC.begin() ;
@@ -79,7 +79,7 @@ int main( int narg, char **argv ) {
 			if ( pos->first < it->second ) break ;
 			pos++ ;
 		}
-		sizeCC.insert( pos, std::pair<arma::u16,int32_t>( it->second, it->first ) ) ;
+		sizeCC.insert( pos, std::pair<arma::u32,int32_t>( it->second, it->first ) ) ;
 	}
 	uint *NewMap = new uint [ /*histogram._bin.size()*/histogram._bin.rbegin()->first+1 ] ;
 	uint iNewMap = 1 ;
@@ -89,7 +89,7 @@ int main( int narg, char **argv ) {
 
 	/// keep only connected components having one of the top_size biggest area value
 
-	arma::u16 last_distinct_value = sizeCC.front().first ;
+	arma::u32 last_distinct_value = sizeCC.front().first ;
 	pos = sizeCC.begin() ;
 	/// move pos to the first item to be removed
 	while ( pos != sizeCC.end() ) {
@@ -115,7 +115,7 @@ int main( int narg, char **argv ) {
 					(*labels)(y,x,z) = NewMap[ (*labels)(y,x,z) ] ;
 	labels->setMaxValue( iNewMap );
 	delete [] NewMap ;
-	IOPgm3d< arma::u16,qint16, false >::write( *labels, output ) ;
+	IOPgm3d< arma::u32,qint32, false >::write( *labels, output ) ;
 	
 	delete labels ;
